Rename seedlist2 to seedlist
