+++ GameCube/include/rando/seedlist.h
